Plot extra information like actual number of substitutions (non, syn, both) per codon site, actual number of unambiguous codons per site.
It appears that even though in the middle of the genome, there is a very high mutation rate, the number of unambigous codons is very low.  This is probably because the reads mapped poorly according to bowtie.  Should look at what happens with a more sensitive alignment and/or more reads.

These results are from the run generated on nibbler with these settings:

--sam "/home/thuy/gitrepo/SlidingWindow/test/simulations/data/sample_genomes.grinder-reads.rename.sam" --ref "consensus" --ref_len 9000 --out_dir "/home/thuy/gitrepo/SlidingWindow/test/simulations/data/out/consensus" --map_q_thresh 10 --read_q_thresh 10 --max_N_thresh 0.1 --window_size 300 --window_breadth_thresh 0.75 --window_depth_thresh 50 --start_nucpos 1 --end_nucpos 9000 --pvalue 0.05 --threads_per_window 3 --concurrent_windows 5 --dnds_tsv "/home/thuy/gitrepo/SlidingWindow/test/simulations/data/out/consensus/actual_dnds_by_site.tsv" --hyphy_exe "/home/thuy/programs/hyphy/hyphy-2.2/HYPHYMP" --hyphy_basedir "/home/thuy/programs/hyphy/hyphy-2.2/res/TemplateBatchFiles/" --fastree_exe "/home/thuy/programs/fasttree/FastTreeMP"